Fixed presumed typo in p_oct_label and fixed size issues in p_EulerInv
In p_oct_label for octant 3 there was a check for the wrong existing indices.
In p_EulerInv there was no pre-allocation of a logical array, hence the EulerInv array would be only the size of the last true element, and since it gets inverted this results in missing values.